+++ src/nodo/application/interfaces/__init__.py
@@ -5,7 +5,7 @@ from nodo.application.interfaces.aggregator_service_registry import (
     IAggregatorServiceRegistry,
 )
 from nodo.application.interfaces.download_repository import IDownloadRepository
-from nodo.application.interfaces.torrent_client import ITorrentClient, TorrentStatus
+from nodo.application.interfaces.torrent_client import ITorrentClient
 from nodo.application.interfaces.user_preferences_repository import (
     IUserPreferencesRepository,
 )
@@ -16,5 +16,4 @@ __all__ = [
     "IDownloadRepository",
     "ITorrentClient",
     "IUserPreferencesRepository",
-    "TorrentStatus",
 ]